Creating MainFunctions
This is the core webrequest for the application.

To be rewritten to call static information from a text document (or several)... eventually.
Note to self: Minimize number of webrequests, optimize code too. 
Example: SystemID to Name/RegionID to Name should pull names from a list rather than submit a search query. Messy.